backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
email-validator==2.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
PyMuPDF==1.24.14
sentence-transformers==3.2.1
faiss-cpu==1.9.0.post1
tiktoken==0.8.0
openai==1.57.3
